pyproject.toml
[project]
name = "chat-agent"
version = "0.1.0"
description = "LangChain Chat Agent with OpenAI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "python-dotenv>=1.0.0",
    "openai>=1.7.2",
    "prompt-toolkit>=3.0.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0"
]

[project.scripts]
chat-agent = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []

